Update scripts and commands: replace execute.sh with pandelosp.sh, add .gitignore entries, and updated clus2json.py
