+++ files/pdi/mycoplasma5.pdi
